+++ routes/index.js
@@ -2,13 +2,13 @@ import { Router } from 'express';
 import pizzas from './pizzas';
 import users from './users';
 import auth from './auth';
-import pages from './pages';
+import json from './jsons'
 
 const router = Router();
 
-router.use('/', pages); 
 router.use('/users', users); 
 router.use('/pizzas', pizzas);
 router.use('/auth', auth);
+router.use('/json', json);
 
 module.exports = router;
